note-app/requirements.txt
fastapi
SQLAlchemy
uvicorn
alembic
python-dotenv
psycopg2-binary
psycopg2
passlib[bcrypt]
python-jose[cryptography]
python-multipart
fastapi-mail